requirements-prod.txt
fastapi>=0.103.2
uvicorn>=0.23.2
gunicorn>=21.2.0
python-decouple>=3.8
pydantic>=2.4.2
pydantic-settings>=2.0.0
email-validator>=2.0.0
python-multipart>=0.0.6
httpx>=0.25.0
prisma>=0.11.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4 